packages/api/pyproject.toml
[project]
name = "lecturelink-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.0.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "dateparser>=1.2.0",
    "python-dateutil>=2.9.0",
    "loguru>=0.7.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "httpx>=0.27.0",
    "numpy>=2.4.2",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-tasks>=2.16.0",
    "redis[hiredis]>=5.0",
    "arq>=0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "tests"]
markers = [
    "integration: Integration tests requiring mocked or real Supabase",
    "slow: Tests that take more than 10 seconds",
    "live: Tests requiring real Gemini API and real Supabase (opt-in)",
]
filterwarnings = ["ignore::DeprecationWarning"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
